remove wp_inverted to make mmc device non-readonly.
with qemu perspetive, it always sets SDHC_WRITE_PROTECT to "prnsts" flag in sdhci.c.
in sd.c, sw_switch flag is set to zero becuae of prnsts flag.
qemu return non-readonly to the kernel.
wp_inverted in the device tree inverts the flag so mmc device is
readonly.

even though I added "read-only=on" as the option of if=sd in qemu commandline,
the sd drive will complain sd can't be readonly device and qemu stopped
running.

it is impossibe to use qemu as the real hw setting.